Pass selectedValue as parameter in MyToggleButton and MyDropdown
